requirements/base.txt
Django==5.1.7
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
python-dotenv==1.0.1
Pillow==10.1.0